requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
boto3==1.41.5
botocore==1.41.5
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
jmespath==1.0.1
joblib==1.5.2
kombu==5.5.4
numpy==2.3.5
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==5.2.1
s3transfer==0.15.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
